refactor: changed Post id to long from integer

+++ src/main/java/com/codecool/poster/controller/PostController.java
@@ -57,7 +57,7 @@ public class PostController {
     }
 
     @GetMapping("/{id}")
-    public Post getPostWithId(@PathVariable Integer id) {
+    public Post getPostWithId(@PathVariable long id) {
         return postService.findById(id);
     }
